+++ bin/check-mysql-select-count.rb
@@ -1,4 +1,6 @@
 #!/usr/bin/env ruby
+# frozen_string_literal: false
+
 #
 # MySQL Select Count Check
 #
@@ -108,6 +110,6 @@ class MysqlSelectCountCheck < Sensu::Plugin::Check::CLI
   rescue StandardError => e
     critical "unhandled exception: #{e}"
   ensure
-    db.close if db
+    db&.close
   end
 end