+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This CHANGELOG follows the format listed [here](https://github.com/sensu-plugins/community/blob/master/HOW_WE_CHANGELOG.md)
 
 ## [Unreleased]
+### Changed
+- `metrics-mysql-graphite.rb`: collect more InnoDB variables. (@boutetnico)
 
 ## [3.1.1] - 2019-03-04
 ### Fixed