+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This CHANGELOG follows the format listed [here](https://github.com/sensu-plugins/community/blob/master/HOW_WE_CHANGELOG.md)
 
 ## [Unreleased]
+- metrics-mysql-graphite.rb change mysql gem to mysql2 to fix runs against MySQL8.
 
 ## [3.1.1] - 2019-03-04
 ### Fixed